+++ packages/user/src/supertokens/plugin.ts
@@ -43,6 +43,35 @@ const plugin = async (
 
   fastify.decorate("verifySession", verifySession);
 
+  // [RL 2024-06-11] change sRefreshToken cookie path from config
+  fastify.addHook("onSend", async (request, reply) => {
+    const refreshTokenCookiePath =
+      request.server.config.user.supertokens.refreshTokenCookiePath;
+
+    const setCookieHeader = reply.getHeader("set-cookie");
+
+    if (setCookieHeader && refreshTokenCookiePath) {
+      const cookies = Array.isArray(setCookieHeader)
+        ? setCookieHeader
+        : [setCookieHeader];
+
+      const updatedCookies = cookies.map((cookie) => {
+        if (String(cookie).startsWith("sRefreshToken")) {
+          return String(cookie).replace(
+            // eslint-disable-next-line unicorn/better-regex
+            /Path=\/[^;]*/i,
+            `Path=${refreshTokenCookiePath}`
+          );
+        }
+
+        return cookie;
+      });
+
+      reply.removeHeader("set-cookie");
+      reply.header("set-cookie", updatedCookies);
+    }
+  });
+
   done();
 };