+++ packages/user/src/userContext.ts
@@ -1,11 +1,9 @@
 import { wrapResponse } from "supertokens-node/framework/fastify";
 import { EmailVerificationClaim } from "supertokens-node/recipe/emailverification";
 import Session from "supertokens-node/recipe/session";
-import UserRoles from "supertokens-node/recipe/userroles";
 
-import getUserService from "./lib/getUserService";
+import ProfileValidationClaim from "./supertokens/utils/profileValidationClaim";
 
-import type { User } from "./types";
 import type { FastifyRequest, FastifyReply } from "fastify";
 import type { MercuriusContext } from "mercurius";
 
@@ -14,49 +12,25 @@ const userContext = async (
   request: FastifyRequest,
   reply: FastifyReply
 ) => {
-  const { config, slonik, dbSchema } = request;
-
-  let userId: string | undefined;
-
   try {
-    const session = await Session.getSession(request, wrapResponse(reply), {
+    request.session = (await Session.getSession(request, wrapResponse(reply), {
       sessionRequired: false,
       overrideGlobalClaimValidators: async (globalValidators) =>
         globalValidators.filter(
           (sessionClaimValidator) =>
-            sessionClaimValidator.id !== EmailVerificationClaim.key
+            ![EmailVerificationClaim.key, ProfileValidationClaim.key].includes(
+              sessionClaimValidator.id
+            )
         ),
-    });
-
-    userId = session === undefined ? undefined : session.getUserId();
+    })) as (typeof request)["session"];
   } catch (error) {
     if (!Session.Error.isErrorFromSuperTokens(error)) {
       throw error;
     }
   }
 
-  if (userId && !context.user) {
-    const service = getUserService(config, slonik, dbSchema);
-
-    /* eslint-disable-next-line unicorn/no-null */
-    let user: User | null = null;
-
-    try {
-      user = await service.findById(userId);
-    } catch {
-      // FIXME [OP 2022-AUG-22] Handle error properly
-      // DataIntegrityError
-    }
-
-    if (!user) {
-      throw new Error("Unable to find user");
-    }
-
-    const { roles } = await UserRoles.getRolesForUser(userId);
-
-    context.user = user;
-    context.roles = roles;
-  }
+  context.user = request.user;
+  context.roles = request.user?.roles;
 };
 
 export default userContext;